⚡️ 프로필타입 분리
