+++ pages/wiki/[code].tsx
@@ -1,34 +1,14 @@
 import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
+import { ProfileAnswer } from 'types/profile';
 
 import instance from '@/lib/axios-client';
 
 import Contents from './components/Contents';
 
-export interface Profile {
-  updatedAt: string;
-  securityAnswer: string;
-  securityQuestion: string;
-  teamId: string;
-  content: string;
-  nationality: string;
-  family: string;
-  bloodType: string;
-  nickname: string;
-  birthday: string;
-  sns: string;
-  job: string;
-  mbti: string;
-  city: string;
-  image: string;
-  code: string;
-  name: string;
-  id: number;
-}
-
-const getProfileData = async (code: string): Promise<Profile | null> => {
+const getProfileData = async (code: string): Promise<ProfileAnswer | null> => {
   try {
-    const res = await instance.get<Profile>(`/profiles/${code}`);
+    const res = await instance.get<ProfileAnswer>(`/profiles/${code}`);
     return res.data;
   } catch (e) {
     console.error('프로필을 불러오지 못했습니다.', e);
@@ -37,7 +17,7 @@ const getProfileData = async (code: string): Promise<Profile | null> => {
 };
 
 export default function Wiki() {
-  const [profile, setProfile] = useState<Profile | null>(null);
+  const [profile, setProfile] = useState<ProfileAnswer | null>(null);
   const router = useRouter();
   const { code } = router.query;
